(libchdr) Resync against https://github.com/rtissera/libchdr/commit/c05c91300a672b96e077ddde144b3737a01639d8#diff-b4791a43102472fe8d3959f86bd376c4

+++ libretro-common/formats/libchdr/libchdr_lzma.c
@@ -88,7 +88,7 @@ static void *lzma_fast_alloc(void *p, size_t size)
 	}
 
 	/* alloc a new one and put it into the list */
-	addr = (uint32_t *)malloc(sizeof(uint32_t) * (size + sizeof(uint32_t)));
+	addr = (uint32_t *)malloc(sizeof(uint32_t) * size + sizeof(uintptr_t));
 	if (!addr)
 		return NULL;
